feat: simplify helloworld sample
