+++ helloworld/README.md
@@ -4,39 +4,9 @@ This is **Triple** helloworld example to help you finish a basic RPC invocation 
 
 ## Prerequisites
 
-### install protoc
+### install protocol buffer compiler, protoc, [version3](https://protobuf.dev/programming-guides/proto3/)
 
-#### install with package manager
-
-```shell
-# for Linux, use apt or apt-get
-apt install -y protobuf-compiler
-protoc --version # make sure that version is 3+
-
-# for Macos, using homebrew
-brew install protobuf
-protoc --version # make sure that version is 3+
-```
-
-#### install pre-compiled binary
-
-```shell
-# fetch the pre-compiled protoc corresponding to your operating system and computer architecture
-# protoc-<version>-<os>-<arch>.zip
-# or you can download from github.com/protocolbuffers/protobuf/releases manually
-PB_REL="https://github.com/protocolbuffers/protobuf/releases"
-curl -LO $PB_REL/download/v24.4/protoc-24.4-linux-x86_64.zip
-
-# unzip downloaded file under a directory
-unzip protoc-24.4-linux-x86_64.zip -d $HOME/.local
-
-# update path variable
-export PATH="$PATH:$HOME/.local/bin"
-```
-
-#### build protoc from sources
-
-Please see [**Download Protocol Buffers**](https://protobuf.dev/downloads/).
+Please refer to [**Protocol Buffer Compiler Installation**](https://grpc.io/docs/protoc-installation/) and [**Download Protocol Buffers**](https://protobuf.dev/downloads/).
 
 ### install protoc-gen-go
 
@@ -93,107 +63,22 @@ protoc --go_out=. --go_opt=paths=source_relative --triple_out=. --triple_opt=pat
 
 ### client
 
-```shell
-cd ~/triple_helloworld
-mkdir -p go-client/cmd && cd ./go-client/cmd
-```
-
-Finish **client.go** and put it in **go-client/cmd** directory.
-
-```go
-package main
-
-import (
-	"context"
-	"dubbo.apache.org/dubbo-go/v3/client"
-	// important, must import this for dubbo-go extensions
-	_ "dubbo.apache.org/dubbo-go/v3/imports"
-	greet "triple_helloworld/proto"
-	"triple_helloworld/proto/greettriple"
-	"github.com/dubbogo/gost/log/logger"
-)
-
-func main() {
-	// initialize a Client which is responsible for invoking a certain service. it uses Triple protocol by default
-	// if you want to invoke another service, please initialize a new one
-	cli, err := client.NewClient(
-		// specify target server URL
-		client.WithURL("127.0.0.1:20000"),
-	)
-	if err != nil {
-		panic(err)
-	}
-	
-	svc, err := greettriple.NewGreetService(cli)
-	if err != nil {
-		panic(err)
-	}
-
-	resp, err := svc.Greet(context.Background(), &greet.GreetRequest{Name: "hello world"})
-	if err != nil {
-		logger.Error(err)
-	}
-	logger.Infof("Greet response: %s", resp.Greeting)
-}
-```
+Finish **client.go**. For specific code, please refer to [**this**](https://github.com/apache/dubbo-go-samples/blob/new-triple-samples/helloworld/client.go)
 
 ### server
-```shell
-cd ~/triple_helloworld
-mkdir -p go-server/cmd
-mkdir -p go-server/handler
-```
 
-Implement **GreetService** Interface and put it in **go-server/handler** directory.
-Please refer to [**concrete sample**](https://github.com/apache/dubbo-go-samples/tree/new-triple-samples/helloworld/go-server/handler).
-
-Finish **server.go** and put it in **go-server/cmd** directory.
-
-```go
-package main
-
-import (
-    // important, must import this for dubbo-go extensions
-	_ "dubbo.apache.org/dubbo-go/v3/imports"
-	"dubbo.apache.org/dubbo-go/v3/protocol"
-	"dubbo.apache.org/dubbo-go/v3/server"
-	"triple_helloworld/go-server/handler"
-	"triple_helloworld/proto/greettriple"
-	"github.com/dubbogo/gost/log/logger"
-)
-
-func main() {
-	// initialize a Server for serving multiple services 
-	srv, err := server.NewServer(
-		// use Triple protocol by default
-		server.WithServerProtocol(
-			// specify port to listen on
-			protocol.WithPort(20000),
-		),
-	)
-	if err != nil {
-		panic(err)
-	}
-	// register a certain service
-	if err := greettriple.RegisterGreetServiceHandler(srv, &handler.GreetTripleServer{}); err != nil {
-		panic(err)
-	}
-	if err := srv.Serve(); err != nil {
-		logger.Error(err)
-	}
-}
-```
+Implement **GreetService** Interface and finish **server.go**. For specific code, please refer to [**this**](https://github.com/apache/dubbo-go-samples/blob/new-triple-samples/helloworld/server.go).
 
 ## Build and run
 
 ```shell
-cd ~/triple_helloworld/go-server/cmd
-go build -o server .
+cd ~/triple_helloworld
+go build -o server ./server.go
 ./server
 ```
 
 ```shell
-cd ~/triple_helloworld/go-client/cmd
-go build -o client .
+cd ~/triple_helloworld
+go build -o client ./client.go
 ./client
 ```
