--- registry/zookeeper_new/go-client/cmd/client.go
@@ -1,72 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package main
-
-import (
-	"context"
-	"dubbo.apache.org/dubbo-go/v3"
-	"dubbo.apache.org/dubbo-go/v3/client"
-	_ "dubbo.apache.org/dubbo-go/v3/imports"
-	"dubbo.apache.org/dubbo-go/v3/registry"
-	greet "github.com/apache/dubbo-go-samples/helloworld/proto"
-	"github.com/apache/dubbo-go-samples/helloworld/proto/greettriple"
-	"github.com/dubbogo/gost/log/logger"
-	"time"
-)
-
-func main() {
-	ins, err := dubbo.NewInstance(
-		dubbo.WithRegistry(
-			registry.WithZookeeper(),
-			registry.WithAddress("127.0.0.1:2181"),
-			registry.WithTimeout(3*time.Second),
-			registry.WithGroup("myGroup"),
-			registry.WithRegisterInterface(),
-		),
-	)
-	cli, err := ins.NewClient()
-	if err != nil {
-		panic(err)
-	}
-	gvCli, err := ins.NewClient(
-		client.WithGroup("myInterfaceGroup"),
-		client.WithVersion("myInterfaceVersion"),
-	)
-	if err != nil {
-		panic(err)
-	}
-	svc, err := greettriple.NewGreetService(cli)
-	if err != nil {
-		panic(err)
-	}
-	gvSvc, err := greettriple.NewGreetService(gvCli)
-	if err != nil {
-		panic(err)
-	}
-
-	resp, err := svc.Greet(context.Background(), &greet.GreetRequest{Name: "zookeeper"})
-	if err != nil {
-		logger.Error(err)
-	}
-	logger.Infof("Greet response: %s", resp.Greeting)
-	gvResp, err := gvSvc.Greet(context.Background(), &greet.GreetRequest{Name: "zookeeper"})
-	if err != nil {
-		logger.Error(err)
-	}
-	logger.Infof("Group/Version Greet response: %s", gvResp.Greeting)
-}
