--- registry/zookeeper_new/proto/greet.proto
@@ -1,44 +1,0 @@
-syntax = "proto3";
-
-package greet;
-
-option go_package = "github.com/apache/dubbo-go-samples/registry/zookeeper_new/proto;greet";
-
-message GreetRequest {
-  string name = 1;
-}
-
-message GreetResponse {
-  string greeting = 1;
-}
-
-message GreetStreamRequest {
-  string name = 1;
-}
-
-message GreetStreamResponse {
-  string greeting = 1;
-}
-
-message GreetClientStreamRequest {
-  string name = 1;
-}
-
-message GreetClientStreamResponse {
-  string greeting = 1;
-}
-
-message GreetServerStreamRequest {
-  string name = 1;
-}
-
-message GreetServerStreamResponse {
-  string greeting = 1;
-}
-
-service GreetService {
-  rpc Greet(GreetRequest) returns (GreetResponse) {}
-  rpc GreetStream(stream GreetStreamRequest) returns (stream GreetStreamResponse) {}
-  rpc GreetClientStream(stream GreetClientStreamRequest) returns (GreetClientStreamResponse) {}
-  rpc GreetServerStream(GreetServerStreamRequest) returns (stream GreetServerStreamResponse) {}
-}
